pyproject.toml
[project]
name = "anime-pipeline"
version = "0.1.0"
description = "Local AI anime production pipeline CLI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.16.0",
    "pydantic>=2.11.0",
    "jinja2>=3.1.0",
    "tenacity>=9.1.0",
    "pyyaml>=6.0.2",
    "edge-tts>=7.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
]

[project.scripts]
anime-tool = "anime_pipeline.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/anime_pipeline"]

[tool.hatch.build]
include = [
    "src/anime_pipeline/presentation/preview/templates/*.j2",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
